civ-8888 - compare all values in hearingDaySchedule list - update tests - move hearing day model to commons
