+++ src/test/java/uk/gov/hmcts/reform/civil/handler/tasks/AutomatedHearingNoticeHandlerTest.java
@@ -24,6 +24,7 @@ import uk.gov.hmcts.reform.hmc.model.hearing.HearingRequestDetails;
 import uk.gov.hmcts.reform.hmc.model.hearing.HearingResponse;
 import uk.gov.hmcts.reform.hmc.model.hearing.ListAssistCaseStatus;
 import uk.gov.hmcts.reform.hmc.model.hearing.PartyDetailsModel;
+import uk.gov.hmcts.reform.hmc.model.unnotifiedhearings.HearingDay;
 import uk.gov.hmcts.reform.hmc.model.unnotifiedhearings.PartiesNotified;
 import uk.gov.hmcts.reform.hmc.model.unnotifiedhearings.PartiesNotifiedResponse;
 import uk.gov.hmcts.reform.hmc.model.unnotifiedhearings.PartiesNotifiedResponses;
@@ -325,7 +326,9 @@ class AutomatedHearingNoticeHandlerTest {
                 PartiesNotifiedResponse.builder()
                     .serviceData(PartiesNotifiedServiceData
                                      .builder()
-                                     .hearingDate(HEARING_DATE)
+                                     .days(List.of(HearingDay.builder()
+                                                       .hearingStartDateTime(HEARING_DATE)
+                                                       .build()))
                                      .hearingLocation("111111")
                                      .build())
                     .build()
@@ -371,7 +374,11 @@ class AutomatedHearingNoticeHandlerTest {
                 PartiesNotifiedResponse.builder()
                     .serviceData(PartiesNotifiedServiceData
                                      .builder()
-                                     .hearingDate(LocalDateTime.of(2030, 6, 6, 12, 0, 0))
+                                     .days(List.of(
+                                         HearingDay.builder()
+                                             .hearingStartDateTime(LocalDateTime.of(2030, 6, 6, 12, 0, 0))
+                                             .build()
+                                     ))
                                      .hearingLocation(VENUE_ID)
                                      .build())
                     .build()
@@ -412,7 +419,11 @@ class AutomatedHearingNoticeHandlerTest {
                     .responseReceivedDateTime(LocalDateTime.now())
                     .serviceData(PartiesNotifiedServiceData
                                      .builder()
-                                     .hearingDate(HEARING_DATE)
+                                     .days(List.of(
+                                         HearingDay.builder()
+                                             .hearingStartDateTime(HEARING_DATE)
+                                             .build()
+                                     ))
                                      .hearingLocation(VENUE_ID)
                                      .build())
                     .build()
@@ -427,7 +438,11 @@ class AutomatedHearingNoticeHandlerTest {
             RECEIVED_DATETIME,
             PartiesNotified.builder().serviceData(PartiesNotifiedServiceData.builder()
                                                       .hearingLocation(VENUE_ID)
-                                                      .hearingDate(HEARING_DATE)
+                                                      .days(List.of(
+                                                          HearingDay.builder()
+                                                              .hearingStartDateTime(HEARING_DATE)
+                                                              .build()
+                                                      ))
                                                       .hearingNoticeGenerated(false).build()).build()
         );
         verify(runtimeService).setVariables(PROCESS_INSTANCE_ID, HearingNoticeSchedulerVars.builder()
